# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "FitIn": {
        "properties": {
          "target": {
            "enum": [
              "coin",
              "lesion"
            ],
            "title": "Target",
            "type": "string"
          },
          "x": {
            "minimum": 0.0,
            "title": "X",
            "type": "number"
          },
          "y": {
            "minimum": 0.0,
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "target"
        ],
        "title": "FitIn",
        "type": "object"
      },
      "FitOut": {
        "properties": {
          "cx": {
            "title": "Cx",
            "type": "number"
          },
          "cy": {
            "title": "Cy",
            "type": "number"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "outline": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outline"
          },
          "r": {
            "title": "R",
            "type": "number"
          }
        },
        "required": [
          "cx",
          "cy",
          "r",
          "outline",
          "method"
        ],
        "title": "FitOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Personal documentation and measurement aid for skin marks. Not a diagnostic tool.",
    "title": "neVus",
    "version": "snapshot"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/images/{image_id}/fit": {
      "post": {
        "description": "A proposal around a tap, in the photo's upright pixels. Nothing is stored.",
        "operationId": "propose_fit_api_images__image_id__fit_post",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Fit",
        "tags": [
          "measurements"
        ]
      }
    },
    "/api/images/{image_id}/scale": {
      "get": {
        "operationId": "image_scale_api_images__image_id__scale_get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageScaleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Image Scale",
        "tags": [
          "measurements"
        ]
      }
    },
    "/api/images/{image_id}/scale-references": {
      "post": {
        "operationId": "add_scale_reference_api_images__image_id__scale_references_post",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "coin": "#/components/schemas/CoinIn",
                    "manual": "#/components/schemas/ManualIn"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CoinIn"
                  },
                  {
                    "$ref": "#/components/schemas/ManualIn"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScaleReferenceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Scale Reference",
        "tags": [
          "measurements"
        ]
      }
    },
...
    "/api/lesions/{lesion_id}/measurements": {
      "get": {
        "operationId": "lesion_measurements_api_lesions__lesion_id__measurements_get",
        "parameters": [
          {
            "in": "path",
            "name": "lesion_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lesion Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeasurementOut"
                  },
                  "title": "Response Lesion Measurements Api Lesions  Lesion Id  Measurements Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lesion Measurements",
        "tags": [
          "measurements"
        ]
      }
    },
    "/api/lesions/{lesion_id}/measurements.csv": {
      "get": {
        "description": "The series as CSV: one row per visit, millimetres with their standard deviation.",
        "operationId": "lesion_measurements_csv_api_lesions__lesion_id__measurements_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "lesion_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lesion Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lesion Measurements Csv",
        "tags": [
          "measurements"
        ]
      }
    },
...
    "/api/measurements/{measurement_id}": {
      "delete": {
        "operationId": "delete_measurement_api_measurements__measurement_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "measurement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Measurement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Measurement",
        "tags": [
          "measurements"
        ]
      }
    },
...
    "/api/observations/{observation_id}/measurements": {
      "get": {
        "operationId": "observation_measurements_api_observations__observation_id__measurements_get",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeasurementOut"
                  },
                  "title": "Response Observation Measurements Api Observations  Observation Id  Measurements Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Observation Measurements",
        "tags": [
          "measurements"
        ]
      },
      "post": {
        "description": "Submitting a measurement is the confirmation: nothing proposed is ever stored without it.",
        "operationId": "create_measurement_api_observations__observation_id__measurements_post",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Measurement",
        "tags": [
          "measurements"
        ]
      }
    },
    "/api/observations/{observation_id}/measurements/preview": {
      "post": {
        "description": "The same computation as saving, without saving: the interface shows it while the outline moves.",
        "operationId": "preview_measurement_api_observations__observation_id__measurements_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Measurement",
        "tags": [
          "measurements"
        ]
      }
    },
...
  }
}